requirements.txt
dotenv
langchain_groq
langchain-core 
langchain-google-genai 
langgraph 
sentence-transformers 
neo4j 
langchain-community 
numpy